--- src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,15 +1,0 @@
-package roomescape.repository;
-
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.reservation.ReservationTime;
-import roomescape.domain.reservation.Time;
-
-@Repository
-public interface ReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
-
-    List<ReservationTime> findAllByStartAt(Time startAt);
-
-    Boolean existsByStartAt(Time startAt);
-}
